Add better support for throttling exceptions
In cases where requests fail for non-technical reasons,
we often do not want exponential backoff. Notably when
requests are throttled it is more appropriate to wait
a specified amount of time before retrying. Often we
will learn this from a `Retry-After` header.

This change introduces `BackoffAndRetryException`
which can be used to sleep a fixed amount of time
before retrying.