Make @retry work on both sync and async functions, with appropriate typing.
This removes the footgun where @retry could accidentally be used on
async functions where it doesn't do what is expected (namely: the sync
implementation would block rather than yield back to the async runtime).

The new implementation inspects whether it is wrapping an async function
(a 'coroutinefunction', c.f. `inspect.iscoroutinefunction`) and picks
the appropriate implementation based on that check.

Type overloading is used to propagate type information to the usage site
of the decorated function, also regardless of whether it is sync or async.